Uploaded research data and analysis on Spirulina bioactivity (in vitro & in silico)
This interdisciplinary project investigates the antioxidant and antimicrobial properties of Spirulina platensis using a combination of in vitro assays (DPPH, Fe³⁺ reducing power, etc.) and in silico methods (molecular docking and ADMET profiling). The aim is to identify potential bioactive compounds with therapeutic relevance for microbial infections and oxidative stress.